# project/its-api-docs/spec/openapi.yaml
swagger: "2.0"
info:
  description: API for Issue tracking system
  version: 1.0.0
  title: ITS - API

# tags are used for organizing operations
...
paths:
...
  /profile:
    get:
      tags:
        - profile
      summary: get profile by id
      operationId: profileById
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: query
          type: integer
          name: profileId
          required: true
      responses:
        200:
          description: fetch profile success
          schema:
            $ref: "#/definitions/PROFILE_RES"
        400:
          description: Invalid parameter
        404:
          description: Profile not found
        500:
          description: Internal Server Error
  /profile/me:
    get:
      tags:
        - profile
      summary: get my username
      operationId: profileMe
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: query
          type: string
          name: sessionId
          required: true
        - in: query
          type: integer
          name: profileId
          required: true
      responses:
        200:
          description: fetch profile success
          schema:
            $ref: "#/definitions/PROFILE_RES2"
        400:
          description: Invalid parameter
        401:
          description: sessionId unauthorized
        500:
          description: Internal Server Error
  /profile/update:
    put:
      tags:
        - profile
      summary: update my profile
      operationId: profileUpdate
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: query
          type: string
          name: sessionId
          required: true
        - in: body
          name: FormData
          schema:
            $ref: "#/definitions/UpdateUserReq"
      responses:
        200:
          description: fetch profile success
          schema:
            $ref: "#/definitions/PROFILE_RES2"
        400:
          description: Invalid parameter
        401:
          description: sessionId unauthorized
        500:
          description: Internal Server Error
  /profile/checkpassword:
    post:
      tags:
        - profile
      summary: check password correct
      operationId: checkPassword
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: query
          type: string
          name: sessionId
          required: true
        - in: body
          name: password
          schema:
            type: object
            required:
              - password
            properties:
              password:
                type: string
                format: hash
                example: 127e6fbfe24a750e72930c220a8e138275656b8e5d8f48a98c3c92df2caba935
      responses:
        200:
          description: password correct
        400:
          description: Invalid parameter
        401:
          description: sessionId unauthorized
        406:
          description: password not correct
        500:
          description: Internal Server Error
  /profile/username:
    get:
      tags:
        - profile
      summary: get like username
      operationId: searchLikeUsername
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: query
          type: string
          name: sessionId
          required: true
        - in: query
          type: string
          name: username
          required: true
      responses:
        200:
          description: search username success
          schema:
            type: array
            items: 
              type: string
        400:
          description: Invalid parameter
        401:
          description: sessionId unauthorized
        500:
          description: Internal Server Error
  /media/image:
    post:
      tags:
        - media
      summary: Upload image and return url
      operationId: mediaImage
      consumes:
        - multipart/form-data
      produces:
        - application/json
      parameters:
        - in: query
          type: string
          name: sessionId
          required: true
        - in: body
          name: FormData
          schema:
            type: object
            properties:
              file:
                type: object
      responses:
        200:
          description: upload success
          schema: 
             type: string
             format: string
             example: "www.test.imagepath.jpg"
        400:
          description: Invalid parameter
        401:
          description: sessionId unauthorized
        500:
          description: Internal Server Error
...
  /project/getMember:
    post:
      tags:
        - project
      summary: get project member by project id
      operationId: getProjectMember
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: query
          type: string
          name: sessionId
          required: true
        - in: body
          name: FormData
          schema:
            $ref: "#/definitions/GetProjectIssueReq"
      responses:
        200:
          description: fetch member success
          schema:
            type: array
            items: 
              $ref: "#/definitions/ProjectMemberRes"
        400:
          description: Invalid parameter
        401:
          description: sessionId unauthorized
        403:
          description: permission denied
        500:
          description: Internal Server Error
...
  /issue:
    post:
      tags:
        - issue
      summary: New issue
      operationId: newIssue
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: query
          type: string
          name: sessionId
          required: true
        - in: body
          name: FormData
          schema:
            $ref: "#/definitions/NewIssueReq"
      responses:
        200:
          description: fetch profile success
          schema:
            type: array
            items: 
              $ref: "#/definitions/NewIssueRes"
        400:
          description: Invalid parameter
        401:
          description: sessionId unauthorized
        500:
          description: Internal Server Error
    put:
      tags:
        - issue
      summary: Edit issue
      operationId: editIssue
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: query
          type: string
          name: sessionId
          required: true
        - in: body
          name: FormData
          schema:
            $ref: "#/definitions/EditIssueReq"
      responses:
        200:
          description: success
          schema:
            type: array
            items: 
              $ref: "#/definitions/NewIssueRes"
        400:
          description: Invalid parameter
        401:
          description: sessionId unauthorized
        500:
          description: Internal Server Error
  /issue/project:
    post:
      tags:
        - issue
      summary: Get issue by project Id
      operationId: projectIssue
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: query
          type: string
          name: sessionId
          required: true
        - in: body
          name: FormData
          schema:
            $ref: "#/definitions/GetProjectIssueReq"
      responses:
        200:
          description: fetch profile success
          schema:
            type: array
            items: 
              $ref: "#/definitions/NewIssueRes"
        400:
          description: Invalid parameter
        401:
          description: sessionId unauthorized
        500:
          description: Internal Server Error
  /issue/get:
    get:
      tags:
        - issue
      summary: fetch issue by issueId
      operationId: getIssueById
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: query
          type: string
          name: sessionId
          required: true
        - in: query
          type: string
          name: issueId
          required: true
      responses:
        200:
          description: success
          schema:
            type: array
            items: 
              $ref: "#/definitions/IssueDetailRes"
        400:
          description: Invalid parameter
        401:
          description: sessionId unauthorized
        500:
          description: Internal Server Error
    post:
      tags:
        - issue
      summary: fetch issue list
      operationId: getUserIssue
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: query
          type: string
          name: sessionId
          required: true
        - in: body
          name: FormData
          schema:
            $ref: "#/definitions/GetProjectReq"
      responses:
        200:
          description: success
          schema:
            type: array
            items: 
              $ref: "#/definitions/NewIssueRes"
        400:
          description: Invalid parameter
        401:
          description: sessionId unauthorized
        500:
          description: Internal Server Error
  /issue/activity:
    post:
      tags:
        - issue
      summary: new issue activity
      operationId: newIssueActivity
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: query
          type: string
          name: sessionId
          required: true
        - in: body
          name: FormData
          schema:
            $ref: "#/definitions/NewActivityReq"
      responses:
        200:
          description: success
          schema:
            type: array
            items: 
              $ref: "#/definitions/IssueDetailRes"
        400:
          description: Invalid parameter
        401:
          description: sessionId unauthorized
        500:
          description: Internal Server Error
    put:
      tags:
        - issue
      summary: edit issue activity
      operationId: editIssueActivity
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: query
          type: string
          name: sessionId
          required: true
        - in: body
          name: FormData
          schema:
            $ref: "#/definitions/EditActivityReq"
      responses:
        200:
          description: success
          schema:
            type: array
            items: 
              $ref: "#/definitions/IssueDetailRes"
        400:
          description: Invalid parameter
        401:
          description: sessionId unauthorized
        500:
          description: Internal Server Error
definitions:
...
  PROFILE_RES:
    type: object
    required:
      - username
    properties:
      username:
        type: string
        example: ting
  PROFILE_RES2:
    type: object
    required:
      - username
      - photoUrl
      - lineId
      - isRevice
    properties:
      username:
        type: string
        example: ting
      photoUrl:
        type: string
        example: www.image.jpg
      lineId:
        type: string
        example: ewuhrriudhgisnrhgiueni
      isRevice:
        type: boolean
        example: true
...
  GetProjectReq:
    type: object
    required:
      - user_id      
      - timestamp
      - limit
      - isOwner
    properties:
      user_id:
        type: integer
        example: 15
      timestamp:
        type: string
        example: 2019-12-05T06:40:43.746+0000
      limit:
        type: integer
        example: 15
      isOwner:
        type: boolean
        example: true
...
  NewIssueReq:
    type: object
    properties:
      name:
        type: string
        example: new issue
      description:
        type: string
        example: hello world
      projectId:
        type: integer
        example: 1
      state:
        type: integer
        example: 1
      severity:
        type: integer
        example: 1
      priority:
        type: integer
        example: 1
      isReproducible:
        type: boolean
        example: true
      tags:
        type: array
        items:
          type: string
  EditIssueReq:
    type: object
    properties:
      name:
        type: string
        example: new issue
      description:
        type: string
        example: hello world
      issueId:
        type: integer
        example: 1
      tags:
        type: array
        items:
          type: string
      mentionList:
        type: array
        items:
          type: string
  NewIssueRes:
    type: object
    properties:
      id:
        type: integer
        example: 1
      name:
        type: string
        example: new issue
      description:
        type: string
        example: hello world
      projectId:
        type: integer
        example: 1
      projectName:
        type: string
        example: yo
      projectOwner:
        type: string
        example: Ting
      state:
        type: integer
        example: 1
      severity:
        type: integer
        example: 1
      priority:
        type: integer
        example: 1
      isReproducible:
        type: boolean
        example: true 
      create_time:
        type: string
        example: 2019-12-05T06:40:43.746+0000
      reportUser:
        type: string
        example: test
      reportUserAvatar:
        type: string
        example: www.dsiufhisuhdfisu.jpg
      tags:
        type: array
        items:
          type: string
      assignList:
        type: array
        items:
          type: string
  IssueDetailRes:
    type: object
    properties:
      id:
        type: integer
        example: 1
      name:
        type: string
        example: new issue
      description:
        type: string
        example: hello world
      projectId:
        type: integer
        example: 1
      projectName:
        type: string
        example: yo
      projectOwner:
        type: string
        example: Ting
      state:
        type: integer
        example: 1
      severity:
        type: integer
        example: 1
      priority:
        type: integer
        example: 1
      isReproducible:
        type: boolean
        example: true 
      create_time:
        type: string
        example: 2019-12-05T06:40:43.746+0000
      reportUser:
        type: string
        example: test
      reportUserAvatar:
        type: string
        example: www.dsiufhisuhdfisu.jpg
      tags:
        type: array
        items:
          type: string
      projectIdentity:
        type: integer
        example: 1
      assignList:
        type: array
        items:
          type: string
      mentionList:
        type: array
        items:
          type: string
      activity:
        type: array
        items:
          $ref: "#/definitions/IssueActivityRes"
  GetProjectIssueReq:
    type: object
    properties:
      projectId:
        type: integer
        example: 2
      timestamp:
        type: string
        example: 2019-12-05T06:40:43.746+0000
      limit:
        type: integer
        example: 15
...
  ProjectMemberRes:
    type: object
    properties:
      id:
        type: integer
        example: 2
      username:
        type: string
        example: test
      join_time:
        type: string
        example: 2019-12-05T06:40:43.746+0000
      identity:
        type: integer
        example: 1
      avatarUrl:
        type: string
        example: dawbjkdnawkjndkaw
  UpdateUserReq:
    type: object
    properties:
      avatarUrl:
        type: string
        example: dfgdfgsdfew
      lineId:
        type: string
        example: test
      username:
        type: string
        example: asd
      isRevice:
        type: boolean
        example: true
      password:
        type: string
...
  IssueActivityRes:
    type: object
    properties:
      id:
        type: integer
      type:
        type: integer
      issueId:
        type: integer
      userId:
        type: integer
      create_time:
        type: string
      prevState:
        type: integer
      nextState:
        type: integer
      username:
        type: string
      content:
        type: string
      avatarUrl:
        type: string
      assignList:
        type: array
        items:
          type: string
  NewActivityReq:
    type: object
    properties:
      issueId:
        type: integer
      state:
        type: integer
      severity:
        type: integer
      priority:
        type: integer
      isReproducible:
        type: boolean
      content:
        type: string
      mentionList:
        type: array
        items:
          type: string
      assignList:
        type: array
        items:
          type: string
  EditActivityReq:
    type: object
    properties:
      activityId:
        type: integer
      issueId:
        type: integer
      name:
        type: string
      comment:
        type: string
      mentionList:
        type: array
        items:
          type: string