# project/its-api/src/main/java/com/its/itsapi/model/request/profile_req/UpdateUserReq.java
package com.its.itsapi.model.request.profile_req;

public class UpdateUserReq {
    public String avatarUrl;
    public String lineId;
    public String username;
    public Boolean isRevice;
    public String password;
}